requirements.txt
# TRACE - Traffic & Resource Agentic Control Engine
# Python Dependencies

# Google Agent Development Kit
google-adk>=0.1.0

# AWS SDK for Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Model Context Protocol
mcp>=0.1.0

# Async/Concurrency
aiohttp>=3.9.0
asyncio

# Monitoring & Telemetry
prometheus-client>=0.19.0
psutil>=5.9.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
pydantic>=2.5.0
